Manage configuration reload from the extension
Placing a new Extension class in the extension module allows us to have
simpler control over the plasticscm object and reload it when
configuration changes.

Some configuration changes can be updated on the fly in the extension.
Others, such as the 'cm' path, need to stop it completely and start it again.

Signed-off-by: Miguel González <mgonzalez@codicesoftware.com>