Use the wk info as root for all shell ops
As the workspace class models a full Plastic SCM workspace,
we can't use the VS Code folder as root to run operations. If some user
has two different folders in the same Plastic SCM workspace opened at
the same time in VS Code, we need to fetch changes for both.

Signed-off-by: Miguel González <miguel.gonzalez@unity3d.com>